using simple rest with Controller
