--- src/main/java/br/edu/ifrn/conta/restclient/ContaCreditoEmbedded.java
@@ -1,10 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import br.edu.ifrn.conta.domain.ContaCredito;
-
-import lombok.Data;
-
-@Data
-public class ContaCreditoEmbedded {
-    private Iterable<ContaCredito> contasCredito;
-}
