--- src/main/java/br/edu/ifrn/conta/restclient/ContaDebitoDTO.java
@@ -1,8 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import lombok.Data;
-
-@Data
-public class ContaDebitoDTO {
-    private ContaDebitoEmbedded _embedded;
-}
