--- src/main/java/br/edu/ifrn/conta/restclient/DonoEmbedded.java
@@ -1,10 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import br.edu.ifrn.conta.domain.Dono;
-
-import lombok.Data;
-
-@Data
-public class DonoEmbedded {
-    private Iterable<Dono> donos;
-}
