--- src/main/java/br/edu/ifrn/conta/restclient/LancamentoDTO.java
@@ -1,8 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import lombok.Data;
-
-@Data
-public class LancamentoDTO {
-    private LancamentoEmbedded _embedded;
-}
