--- src/main/java/br/edu/ifrn/conta/restclient/LancamentoEmbedded.java
@@ -1,10 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import br.edu.ifrn.conta.domain.Lancamento;
-
-import lombok.Data;
-
-@Data
-public class LancamentoEmbedded {
-    private Iterable<Lancamento> lancamentos;
-}
