+++ src/main/java/br/edu/ifrn/conta/restclient/LancamentoRestClient.java
@@ -1,8 +1,10 @@
 package br.edu.ifrn.conta.restclient;
 
 import br.edu.ifrn.conta.domain.Lancamento;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Data;
+import org.springframework.web.util.UriComponentsBuilder;
 
 @Data
 @Builder
@@ -11,8 +13,7 @@ public class LancamentoRestClient {
     private RestClientHelper<Lancamento> entityRestHelper;
 
     public Iterable<Lancamento> findAll() {
-        return this.entityRestHelper.getRestTemplate().getForObject(
-                entityRestHelper.findAllUri(), LancamentoDTO.class).get_embedded().getLancamentos();
+        return this.entityRestHelper.findAll();
     }
 
     public Lancamento save(Lancamento entity) {
@@ -20,4 +21,13 @@ public class LancamentoRestClient {
                 entityRestHelper.saveUri(), entity, Lancamento.class);
     }
 
+    public BigDecimal saldo(String dono, String contaPatrimonio) {
+        return this.entityRestHelper.getRestTemplate().getForObject(
+        UriComponentsBuilder.fromHttpUrl(entityRestHelper.url() + "/saldo")
+            .queryParam("dono", dono)
+            .queryParam("contaPatrimonio", contaPatrimonio)
+            .build().encode().toUri()
+            , BigDecimal.class);
+    }
+    
 }
