--- src/main/java/br/edu/ifrn/conta/restclient/ValorInicialDoDonoNaContaPatrimonioDTO.java
@@ -1,8 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import lombok.Data;
-
-@Data
-public class ValorInicialDoDonoNaContaPatrimonioDTO {
-    private ValorInicialDoDonoNaContaPatrimonioEmbedded _embedded;
-}
