--- src/main/java/br/edu/ifrn/conta/restclient/ValorInicialDoDonoNaContaPatrimonioEmbedded.java
@@ -1,10 +1,0 @@
-package br.edu.ifrn.conta.restclient;
-
-import br.edu.ifrn.conta.domain.ValorInicialDoDonoNaContaPatrimonio;
-
-import lombok.Data;
-
-@Data
-public class ValorInicialDoDonoNaContaPatrimonioEmbedded {
-    private Iterable<ValorInicialDoDonoNaContaPatrimonio> valoresIniciaisDoDonoNaContaPatrimonio;
-}
